requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
blobfile==2.0.1
cachetools==5.3.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
filelock==3.9.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
googleapis-common-protos==1.58.0
googletrans==4.0.0rc1
grpcio-status==1.51.1
grpcio==1.51.1
h11==0.9
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httptools==0.5.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.8
lxml==4.9.2
multidict==6.0.4
openai==0.26.5
pip==22.3.1
proto-plus==1.22.2
protobuf==4.21.12
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycryptodomex==3.17
pydantic==1.10.4
python-dotenv==0.21.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
setuptools==65.5.1
six==1.16.0
slack-sdk==3.19.5
sniffio==1.3.0
SQLAlchemy==2.0.2
starlette==0.22.0
tiktoken==0.2.0
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wheel==0.38.4
yarl==1.8.2
revChatGPT==2.1.14